requirements.txt
Django==2.2.3
django-cors-headers==3.1.0
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
djangorestframework-jwt
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
mysql-connector
django-mysql
gunicorn
